requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.32.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
flower==2.0.0
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
humanize==4.7.0
idna==3.4
kombu==5.3.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
